requirements.txt
coverage==4.3.4
Flask==1.1.4
Flask-Cors==3.0.10
Flask-OAuth==0.12
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.6.2
gunicorn==19.6.0
nose==1.3.7
flasgger=0.9.0

